# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/AAhbari/Assignment1/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://virtserver.swaggerhub.com/tutorial/1.0.0
  description: SwaggerHub API Auto Mocking
paths:
  /student/{student_id}:
    get:
      summary: gets student
      description: |
        Returns a single student
      operationId: get_student_by_id
      parameters:
      - name: student_id
        in: path
        description: the uid
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        "400":
          description: invalid ID
        "404":
          description: not found
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: deletes student
      description: |
        delete a single student
      operationId: delete_student
      parameters:
      - name: student_id
        in: path
        description: the uid
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: integer
      responses:
        "200":
          description: successful operation
        "400":
          description: invalid ID
        "404":
          description: not found
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /student:
    post:
      summary: adds student
      description: |
        Adds a student to the system
      operationId: add_student
      requestBody:
        description: Student item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        "200":
          description: created
          content:
            application/json:
              schema:
                type: number
                format: integer
                x-content-type: application/json
        "400":
          description: "invalid input, object invalid"
        "409":
          description: an existing already exists
      x-openapi-router-controller: swagger_server.controllers.default_controller
components:
  schemas:
    Student:
      required:
      - first_name
      - last_name
      - student_id
      type: object
      properties:
        student_id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        gradeRecords:
          type: array
          items:
            $ref: '#/components/schemas/GradeRecord'
      example:
        student_id: 0
        last_name: last_name
        first_name: first_name
        gradeRecords:
        - grade: 6
          subject_name: subject_name
        - grade: 6
          subject_name: subject_name
    GradeRecord:
      required:
      - grade
      - subject_name
      type: object
      properties:
        subject_name:
          type: string
        grade:
          type: integer
      example:
        grade: 6
        subject_name: subject_name